+++ Discord/IDiscordWebhook.cs
@@ -1,0 +1,10 @@
+﻿using Discord;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace DanielLudwig.NINA.DiscordNotification.Discord {
+	public interface IDiscordWebhook {
+		Task SendMessage(string text = null, IEnumerable<EmbedFieldBuilder> fields = null);
+		Task SendFileMessage(string filePath, string text = null, IEnumerable<EmbedFieldBuilder> fields = null);
+	}
+}